@@ -1,26 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="IdentityExpress.Identity" Version="2.3.0.12" />
-    <PackageReference Include="IdentityExpress.Manager.Api.Community" Version="2.4.0.133" />
-    <PackageReference Include="IdentityServer4" Version="2.5.0" />
-    <PackageReference Include="IdentityServer4.EntityFramework" Version="2.5.0" />
-    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="2.5.0" />
-    <PackageReference Include="RSK.Audit" Version="1.0.0.35" />
-    <PackageReference Include="RSK.Audit.EF" Version="1.0.0.35" />
-    <PackageReference Include="RSK.IdentityServer4.AuditEventSink" Version="1.0.0.19" />
+    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="3.0.1" />
 
-    <PackageReference Include="Serilog" Version="2.8.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="2.1.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="3.0.0" />
+    
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="3.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.0.0" />
+
+    <PackageReference Include="Serilog" Version="2.9.0" />
+    <PackageReference Include="Serilog.AspNetCore" Version="3.1.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
     <PackageReference Include="Serilog.Sinks.File" Version="4.0.0" />
-    
-    <!-- <PackageReference Include="Microsoft.AspNetCore.App" /> -->
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.6" PrivateAssets="All" />
   </ItemGroup>
 </Project>
