@@ -1,4 +1,4 @@
-@model LoginViewModel
+﻿@model LoginViewModel
 
 <div class="login-page">
     <div class="page-header">
@@ -39,6 +39,11 @@
                                         </label>
                                     </div>
                                 }
+
+                                <div>
+                                    <p>The default users are alice/bob, password: Pass123$</p>
+                                </div>
+
                                 <div class="form-group">
                                     <button class="btn btn-primary" name="button" value="login">Login</button>
                                     <button class="btn btn-default" name="button" value="cancel">Cancel</button>
