@@ -1,4 +1,4 @@
-@model ErrorViewModel
+﻿@model ErrorViewModel
 
 @{
     var error = Model?.Error?.Error;
