@@ -37,7 +37,7 @@ namespace Third.MVC
                 options.Authority = "http://localhost:8000";
                 options.RequireHttpsMetadata = false;
                 options.ClientId = "mvc";
-                options.ClientSecret = "lesson_3";
+                options.ClientSecret = "49C1A7E1-0C79-4A89-A3D6-A37998FB86B0";
                 options.ResponseType = "id_token code";
 
                 options.SaveTokens = true;
